requirements.txt
python-dotenv
pandas
pyarrow
upstox-python-sdk
python-dotenv